CGAME: Restore viewsize functionality and cvar
- Restore viewsize functionality
- Restore cg_viewsize cvar
- Add missing backtile asset
- Should fix #153
- There are still issues with the sky lens flare

+++ code/ui/ui_display.c
@@ -222,7 +222,7 @@ static void UI_DisplayOptions_Event(void *ptr, int event) {
 		break;
 
 	case ID_SCREENSIZE:
-		trap_Cvar_SetValue("cg_viewsize", (float)displayOptionsInfo.screensize.curvalue);
+		trap_Cvar_SetValue("cg_viewsize", (int)displayOptionsInfo.screensize.curvalue);
 		break;
 
 	case ID_MAXFPS:
